Fix internal vector growing in sinkVector.
A slight oversight mas made in that `VM.grow`'s second parameter (`by`) defines the amount by which the vector's size will be incremented. Thus, the new vector's size will be `(VM.length vec) + by`.
As implemented, the function assumed that the `by` parameter defines the total new size. As a result, the vector's size grew very fast although only a small subset was being used.
I have assumed that a geometric progression with a ratio of 2 was the intended operation and this is what my change implements. I have tested it and it works fine.
Thanks, for all the great work!