Remove whitespace after '@' in as-patterns
Currently, `mono-traversable` has these two lines of code:

```hs
    oextend f w@ ~(_ :< xxs) =
```
```hs
    oextend f w@ ~(xxs :> _) =
```

Notice the space after the `@` in each as-pattern. As of GHC proposal
[#229](https://github.com/ghc-proposals/ghc-proposals/blob/master/proposals/0229-whitespace-bang-patterns.rst),
this is a parse error, which means that `mono-traversable` fails to
compile on GHC HEAD:

```
src/Data/MonoTraversable.hs:1317:16: error:
    Suffix occurrence of @. For an as-pattern, remove the leading whitespace.
     |
1317 |     oextend f w@ ~(_ :< xxs) =
     |                ^
```

This is easily fixed in a backwards-compatible way by removing the
whitespace and adding an extra set of parentheses, which this patch
accomplishes.